requirements.txt
kafka-python
pyspark
findspark
pyspark-stubs
pandas